Overlay splitting
