--- Sources/BrowserServicesKit/Common/Extensions/NSViewExtension.swift
@@ -1,34 +1,0 @@
-//
-//  NSViewExtension.swift
-//
-//  Copyright © 2021 DuckDuckGo. All rights reserved.
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//  http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-//
-#if !os(iOS)
-import Cocoa
-
-extension NSView {
-
-    func addAndLayout(_ subView: NSView) {
-        subView.translatesAutoresizingMaskIntoConstraints = false
-        addSubview(subView)
-
-        subView.leadingAnchor.constraint(equalTo: leadingAnchor).isActive = true
-        subView.topAnchor.constraint(equalTo: topAnchor).isActive = true
-        subView.bottomAnchor.constraint(equalTo: bottomAnchor).isActive = true
-        subView.widthAnchor.constraint(equalTo: widthAnchor).isActive = true
-    }
-
-}
-#endif
